// src/interface.ts
export interface asset {
    id: number;
    token_id: string;
    num_sales: number;
    background_color: string;
    image_url: string;
    image_preview_url: string;
    image_thumbnail_url: string;
    image_original_url: string;
    animation_url: string;
    animation_original_url: string;
    name: string;
    description: string;
    external_link: string;
    asset_contract: {
        address: string;
        asset_contract_type: string;
        created_date: string;
        name: string;
        nft_version: string;
        opensea_version: string;
        owner: number;
        schema_name: string;
        symbol: string;
        total_supply: string;
        description: string;
        external_link: string;
        image_url: string;
        default_to_fiat: boolean;
        dev_buyer_fee_basis_points: number;
        dev_seller_fee_basis_points: number;
        only_proxied_transfers: boolean;
        opensea_buyer_fee_basis_points: number;
        opensea_seller_fee_basis_points: number;
        buyer_fee_basis_points: number;
        seller_fee_basis_points: number;
        payout_address: string;
    };
    permalink: string;
    collection: {
        banner_image_url: string;
        chat_url: string;
        created_date: string;
        default_to_fiat: boolean;
        description: string;
        dev_buyer_fee_basis_points: string;
        dev_seller_fee_basis_points: string;
        discord_url: string;
        display_data: {
            card_display_style: string;
        };
        external_url: string;
        featured: boolean;
        featured_image_url: string;
        hidden: boolean;
        safelist_request_status: string;
        image_url: string;
        is_subject_to_whitelist: boolean;
        large_image_url: string;
        medium_username: string;
        name: string;
        only_proxied_transfers: boolean;
        opensea_buyer_fee_basis_points: string;
        opensea_seller_fee_basis_points: string;
        payout_address: string;
        require_email: boolean;
        short_description: string;
        slug: string;
        telegram_url: string;
        twitter_username: string;
        instagram_username: string;
        wiki_url: string;
        is_nsfw: boolean;
    };
    decimals: number;
    token_metadata: string;
    owner: {
        user: {
            username: string;
        };
        profile_img_url: string;
        address: string;
        config: string;
    };
    sell_orders: null;
    creator: {
        user: {
            username: string;
        };
        profile_img_url: string;
        address: string;
        config: string;
    };
    traits: {
        trait_type: string;
        value: string;
        display_type: null;
        max_value: null;
        trait_count: number;
        order: number;
    }[];
};